pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repodesign"
version = "0.1.0"
description = "Codebase-aware multimodal system design planning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "requests>=2.31",
    "PyGithub>=2.1",
    "anthropic>=0.71",
    "openai>=2.6",
    "pyyaml>=6.0",
    "toml>=0.10",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]
training = ["torch>=2.0", "transformers>=4.40", "huggingface-hub>=0.20"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
